VER-80: breaking down and prep for final page (summary)
