+++ src/components/shared.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+export const Button = styled('button')`
+  height: 30px;
+  border-radius: 8px;
+  border: 1px solid #000;
+  background: #fff;
+  margin-right: 3px;
+  box-shadow: 0px 2px 0px 0px #000;
+`
+
+export const Input = styled('input')`
+  width: 100%;
+  height: 35px;
+  margin-bottom: 5px;
+  text-align: center;
+  border-radius: 2px;
+  border: 1px solid var(--black, #000);
+  background: var(--white, #fff);
+`
