+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 
-import App from './pages/App'
+import App from './App'
 import GlobalFonts from './assets/fonts/fonts'
 
 const root = createRoot(document.getElementById('root'))
